Rename gem to Sommelier
The name Stable Marriage Problem and the language used to
describe it (male suitors looking to win the hand of the
ladies) are rooted in a culture that is losing relevance.
I've chosen to rename this gem to modernize the language
and make it more relevant to today's culture.

Sommelier is a French word for a wine steward, known for
being experts in their field of wine. One responsibility of
a sommelier is to help pair a wine with a given dish.